Force footer to start at left edge.
Mixing position: absolute with container-fluid makes for divs that are
left aligned to the left side of the parent container, with the padding
in place, such that it breaks out of the parent container.

By forcing the left edge of the footer to be at the left edge of the
window, width: 100% behaves as we expect, and the existing padding in
container-fluid keeps things suitably spaced from the edges.