Refactor handling of Events in testing and factories
Don't try to guess whether we should create a new event/get an old event
by looking at the connection schema, as this will fail if the schema is
reset without deleting the event - then we end up with more than one event
which we don't want to happen.

This requires changing Factories for models which reference Event to always
get one out of the database. We use a LazyFunction to just get the unique
one from django, thus creating an implicit requirement that there be one
and only one in the queryset.

This also adds a check to make sure EventFactory is not used inside a
transaction, as this causes an error from the database.