Add a setting for contact preferences
Add a nullable boolean preference on the UserInfo regarding contact
preferences.
Add this to the signup form and the profile form on the account page.
Add an extra announcement displayed to users if this field is null
including a link to the account page.
Update the announcement regarding no seat being set to include a link to
the account page.