Fix some issues with the index
"Current Event" is not accurate - it's just the event associated to the domain. "Home" is a bit old-fashioned so I changed it to the event name. This duplicates the event name in the top left which I don't like. However we do the same in episodes.

I would be happy to get rid of this header in both templates.