Don't display errors when clocks seem to mismatch
These appear way more commonly than expected. This change also alters
behaviour so that when the server's clock seems to be behind that of the
client, it only uses the client's calculation of the guess timeout
if the difference is large. This is safe because sending a guess
too early will just get an error message.