Refactor rules to allow non-superuser admins
Move rules predicates for events and teams to teams app, since teams app
defines the concept of the admin team, leaving only the hunts specific
predicates and permission bindings for hunts objects in hunts app.
Update rules predicates to support cases where the object parameter is
None, or is a parent object so they can be used for object type list
views and inline listings.
Update permission bindings and add new rules predicates for new object
types.
Add object permissions model admin mixins from the rules project to all
admins and inlines which should be usable by non super-user admins.
In the case of non-superusers limit the queryset used to populate the
EventAdmin to events for which they're on an admin team
Add unit tests for all rules predicates