Rename CI stages to `test` and `stage2`
Gitlab has added a new static analysis jobs to the SAST job template we
include. Now there are lots of jobs in this template it is impractical
to override their stages.
Rename the `check` stage in our CI to `test` so that these jobs are
included in this stage. Move all `build-app` jobs to `test` stage too to
improve parallelism.
Rename `web-doc-test` stage to `stage2` to avoid implying that the
stages have any kind of logical meaning.