Add support for privacy policy
Adds privacy policy text configuration to the global site configuration
object and renders a acceptance checkbox on the signup for if the
privacy policy is configured.
In order to support privacy policy page on both event and non-event
pages the event base template is adapted to add event content
conditionally on there being an active event.