Fetch release notes from Github with Renovate
Add an encrypted personal access token for an anonymous Github account
to the renovate configuration.
This token is encrypted so that only the renovate bot can read it, is
for an anonymous Github account, and is only authorized to read public
Github repositories, so including it here in the repo is fairly safe.