Simplify permission checking for admin users
Rather than following the long transitive chain of objects back to the
event object for permission checking, just check the user is on an admin
team for the event associated with the active database schema.
Add tests for basic page load for all admin create forms to avoid
regressions here.