Add flag for events without seat assignemnts.
Adds a boolean field to events that says whether or not seat
assignments are a thing. When seat assignments are disabled,
do not show the "please select a seat" announcement and hide
the entire "Details for $event" section of the profile page.

At some point in the future there may be other fields in the
"Details for $event" section, which will require the
profile edit page to be a bit smarter about things.

Seats are still shown in the admin guesses screen, and in
the about page if the authors have them set, but both
are happy with empty values.