Show user profile pictures somewhere. Close #341
Users who aren't admins are still asked to set user profile pictures, but these
don't end up displaying anywhere. This change updates the team views to look
similar to the author view and show users with images rather than just a list.

This also adds a link to the user profile page in the `/admin/teams/<id>` view
so that an admin can see if a user has participated before or if this is their
first time.

Finally, whilst I was there I also made a minor tweak to the profile page to
change it so that rather than showing "None @" it now just shows the hunt name
in hunts that the user did not participate with a team in and a minor tweak to
the authors page to not show a broken image if the author hasn't set a profile
picture.