Fix for building image:web on forks.
This was previously fixed in MR !333 but appears to have gotten lost somewhere
along the way.

For non-forks it works fine without this, becuase it builds/tags with the normal
`hunter2.app/hunter2` name, but when PRing from a fork the stage1 build ends up
tagging images with the fork name (ie, `shanemcc/hunter2`) and thus the nginx
FROM fails because app.tar has a tagged image under a different name.

If we use `${APP_IMAGE_TAG}` instead then it pulls the right image (as that's the
same var used when tagging it during `image:app` and thus within app.tar.