Don't run before_script on included jobs
At the moment we define a global before_script, which ends up applied to
included jobs such as the Gitlab SAST jobs. This fails since the
dependencies for our before_script are not included in that image.
Define the before_script in a base job which is explicitly included in
our jobs instead.