Disable Node.JS security scanner in CI
Gitlab determines that we are a Node.JS application by the existence of
the package.json and finds a load of misleading findings based on the
assumption that our Django templates are actually Squirrelly.JS
templates.
Override the rules for the `nodejs-scan-sast` so that it does not run
for us.