Make renovate pin and update python dependencies
Pin all dependency versions so that renovate makes PRs for updates to
all of our direct dependencies. Configure digest pinning for docker
images since image tags are mutable.
Set lockFileMaintenance so that we get periodic updates (weekly) for
transitive dependencies which are only referenced in the lockfile.
Set unpublishSafe so that renovate will wait for NPM packages to not be
unpublishable before adopting them.
Configure auto-merge for patch, pin, digest and lockFileMaintenance
updates if they pass CI since we usually accept updates like this
without much scrutiny and the commit history will allow us to identify
what went wrong and roll back if this introduces breakage.