Ensure prefetches in admin progress view
Remove the `order_by` from the query in `Episode.get_relative_id` since
it is the same as the default model ordering and this made the query
miss any `prefetch_related` done beforehand.
Remove episode filtering from admin progress view to simplify
prefetching/selecting.
Add more pre-selection and pre-fetching to the queries for the admin
progress view to prevent query count scaling with number of puzzles or
teams.