Remove encrypted GitHub token from renovate config
We can no longer use the free hosted renovate bot, so we are running our
own via a private GitLab scheduled pipeline. This bot doesn't have
access to the hosted bot's private key to decrypt the GitHub token in
our config and is configured with a GitHub API token in its
configuration anyway so doesn't need one specified in the config.