Fix issues with accessing player stats view
Pass the user profile, rather than the user object, to the player stats
generators, since this is the object which implements the
`get_display_name()` method.
Prevent admins from accessing player stats views before the event is
over, since accessing these causes content to be cached in redis and we
don't want admins to accidentally cause early stats to be cached.
Given that no users are permitted to access the views ahead of time
using HTTP 403 for the error doesn't make sense so change this response
to a 404.