Add basic integration options for a discord bot
Add configuration to the Event object for a Discord URL and a Discord
Bot ID.
Add site-wide API tokens which can be used to authenticate the bot to
bot APIs.
Attach a secret token to the Team object and, if the discord parameters
are present on the Event, render information on the team management page
about how to join your team channels through the bot.
Create a team info endpoint which the bot can query to resolve a secret
team token into the team name.