Improve responses for changing passwords
Move the rendering of announcements to a new Vue component
`<alert-list>` and add Django messages to this with the dismissable
attribute. This will show messages about login, logout and password
change to users after the operation completes.
Change the redirect for successful password change to go back to the
full account page rather than to the specific password reset form page.