Fix problems with progress graphs
Use the TeamPuzzleData object as the root of the query for progress to
ensure only one progress time is found per team/puzzle.
Adjust the start time for each puzzle if the team progress time for the
same puzzle is earlier to counteract the effect when a member joins a
team when they already have an earlier solve time than the team's start
time.